handle empty row condition in data_ingest. remove uuid
adjust indices in data_ingest record validation to account for removal of uuid

adjust array index in data_ingest validaterecord

refactor data_ingest validaterecord to use hash keys instead of array indices

Workaround for broken production installs in v1.1.2 of geoblacklight gem

Installing in RAILS_ENV=production mode via "bundle install --without
'development test'" is broken in v1.1.[0-2] of the geoblacklight gem.
This was reported in issue 494 of the GeoBlacklight GitHub project and
fixed in pull request 495.

Until an updated version of the gem is released, one way to work around
the problem with v1.1.2 is to install the fixed version from GitHub
instead of from rubygems.org.

remove uuid. not needed