Add log rotation to standard logfile
If the option to log to Graylog is not chosen then the default is to
log to the standard Rails log file.  This file, by default, uses no
rotation and thus grows unbounded in size.  This change enables
options on the standard Rails logfile to rotate it every 10 MiB and
to keep the last 12 logfiles.

Signed-off-by: Paul Mather <paul.mather@vt.edu>